+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [1.1.2] - TODO
 
 ### Changed
+- Reset player position when under a certain treshold.
 - Fix player position not being reset after loading.